% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getData.R
\name{.authenticate}
\alias{.authenticate}
\title{GUI for getting username and password from user}
\usage{
.authenticate()
}
\value{
No return value
}
\description{
GUI for getting username and password from user
}
