% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{durumDaily}
\alias{durumDaily}
\title{durumDaily}
\format{
The data includes the site unique identifier and daily data for 4 climatic variables (tmin, tmax, precipitation and relative humidity)
}
\usage{
data(durumDaily)
}
\description{
200 sites from durum wheat collection and their daily climatic data.
}
\examples{
if(interactive()){
 # Load durum wheat data with their daily climatic variables obtained from ICARDA database
 data(durumDaily)
}
}
\keyword{datasets}
