% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{durumWC}
\alias{durumWC}
\title{durumWC}
\format{
The data includes the site unique identifier, longitude, latitude and 55 worldclim data \href{https://www.worldclim.org}{worldclim}
}
\usage{
data(durumWC)
}
\description{
200 sites from durum wheat collection and their world clim data.
}
\examples{
if(interactive()){
 # Load durum wheat data with world climatic variables obtained from WorldClim database
 data(durumWC)
 }
}
\keyword{datasets}
