% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{DurumWheatDHEWC}
\alias{DurumWheatDHEWC}
\title{DurumWheatDHEWC}
\format{
More details about the climate data can be found here \href{https://www.worldclim.org/data/worldclim21.html#bio_10m}{worldclim}
}
\usage{
data("DurumWheatDHEWC")
}
\description{
193 samples of Days to heading from durum wheat collection and 55 worldclim data.
}
\examples{
if(interactive()){
 # Load durum wheat DHE data with world climatic variables obtained from WorldClim database
 data("DurumWheatDHEWC")
 }
}
\keyword{datasets}
