% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{FIGS}
\alias{FIGS}
\title{FIGS subset for wheat sodicity resistance}
\format{
A data frame with 201 rows and 15 variables
}
\usage{
data(FIGS)
}
\description{
FIGS subset for wheat sodicity resistance
constructed using the harmonized world soil database HWSD
}
\examples{
if(interactive()){
 data(FIGS)
}
}
\references{
\href{http://www.fao.org/3/aq361e/aq361e.pdf}{HWSD}
}
\keyword{datasets}
