% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getData.R
\name{getAccessions}
\alias{getAccessions}
\title{Getting Accession Passport Data by Crop or Accession Numbers}
\usage{
getAccessions(
  crop = "",
  ori = NULL,
  IG = NULL,
  doi = lifecycle::deprecated(),
  taxon = lifecycle::deprecated(),
  collectionYear = lifecycle::deprecated(),
  coor = FALSE,
  available = FALSE,
  other_id = FALSE
)
}
\arguments{
\item{crop}{Crop name.}

\item{ori}{Country of origin (ISO 3166-1 alpha-3 code).}

\item{IG}{List of accession numbers.}

\item{doi}{\code{[Deprecated]} No longer used as of version 2.0.0.}

\item{taxon}{\code{[Deprecated]} No longer used as of version 2.0.0.}

\item{collectionYear}{\code{[Deprecated]} No longer used as of version 2.0.0.}

\item{coor}{If \code{TRUE}, returns only georeferenced accessions.}

\item{available}{If \code{TRUE}, returns only available accessions for distribution.}

\item{other_id}{If \code{TRUE}, returns other IDs associated with accessions.}
}
\value{
Data frame of accession passport data by crop or accession numbers.
}
\description{
Return a data frame containing passport data of accessions filtered by crop name or accession numbers.
}
\details{
Available crops can be retrieved using \code{\link[icardaFIGSr]{getCrops}}.
}
\examples{
\dontrun{
 # Obtain accession data for durum wheat
 durum <- getAccessions(crop = 'Durum wheat', coor = TRUE)
 }
}
\author{
Khadija Aouzal, Amal Ibnelhobyb, Zakaria Kehel, Fawzy Nawar
}
