% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCrops.R, R/getData.R
\name{getCrops}
\alias{getCrops}
\title{Crops Available in ICARDA's Genebank}
\usage{
getCrops()

getCrops()
}
\value{
A list containing all crops available in ICARDA's Genebank Documentation System.

A data frame with crop codes and names.
}
\description{
this function allows to obtain a list of crops available in ICARDA's Genebank Database, it returns a list with codes and names of available crops.

Retrieve a list of available crops in ICARDA's Genebank.
}
\details{
The crop codes and names are fetched from ICARDA's Genebank database.
}
\examples{
\dontrun{
 # Get list of available crops
 crops <- getCrops()
 }
\dontrun{
 # Get list of available crops
 crops <- getCrops()
 }
}
\author{
Zakaria Kehel, Fawzy Nawar

Khadija Aouzal, Zakaria Kehel, Fawzy Nawar
}
