% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCropTraits.R
\name{getCropTraits}
\alias{getCropTraits}
\title{Getting Traits Associated with Crops from the ICARDA's Genebank Documentation System}
\usage{
getCropTraits(crop)
}
\arguments{
\item{crop}{character. Crop for which to get available traits.}
}
\value{
A data frame with traits that are associated with the crop specified in \code{crop}.
}
\description{
Return a data frame containing traits associated with a particular crop, their description and related identifiers.
}
\details{
\code{getCropTraits} returns a data frame of traits together with their IDs and coding system used for each trait.

Possible inputs for \code{crop} include:
\describe{
\item{'ICAG'}{Aegilops}
\item{'ICB'}{Barley}
\item{'ICBW'}{Bread wheat}
\item{'ILC'}{Chickpea}
\item{'ICDW'}{Durum wheat}
\item{'ILB'}{Faba bean}
\item{'BPL'}{Faba bean BPL}
\item{'IFMI'}{Forage and range}
\item{'IFLA'}{Lathyrus}
\item{'ILL'}{Lentil}
\item{'IFMA'}{Medicago annual}
\item{'IC'}{Not mandate cereals}
\item{'IFPI'}{Pisum}
\item{'ICPW'}{Primitive wheat}
\item{'IFTR'}{Trifolium}
\item{'IFVI'}{Vicia}
\item{'ICWH'}{Wheat hybrids}
\item{'ICWW'}{Wheat wild relatives}
\item{'ILWC'}{Wild Cicer}
\item{'ICWB'}{Wild Hordeum}
\item{'ILWL'}{Wild Lens}
\item{'ICWT'}{Wild Triticum}
}

A list of available crops to use as input for \code{crop} can also be obtained from ICARDA's online server using \code{\link[icardaFIGSr]{getCrops}}.
}
\examples{
\dontrun{
 # Get traits for bread wheat
 breadTraits <- getCropTraits(crop = 'Bread wheat')
}
}
\author{
Khadija Aouzal, Amal Ibnelhobyb, Zakaria Kehel, Fawzy Nawar, Chafik Analy
}
