% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getData.R
\name{getTraits}
\alias{getTraits}
\title{Getting Traits Associated with Crops from the ICARDA's Genebank Database}
\usage{
getTraits(crop)
}
\arguments{
\item{crop}{character. Crop for which to get available traits.}
}
\value{
A data frame with traits that are associated with the crop specified in \code{crop}.
}
\description{
Return a data frame containing traits associated with a particular crop, their description and related identifiers.
}
\details{
\code{getTraits} returns a data frame of traits together with their IDs and descriptions.
The list of available crops to use as input for \code{crop} can be fetched from ICARDA's Genebank database using \code{\link[icardaFIGSr]{getCrops}}.
}
\examples{
\dontrun{
if(interactive()){
 # Get traits for bread wheat
 breadTraits <- getTraits(crop = 'Bread wheat')
 }
}
}
\author{
Khadija Aouzal, Amal Ibnelhobyb, Zakaria Kehel, Fawzy Nawar
}
