% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_prediction.R
\name{make_prediction}
\alias{make_prediction}
\title{Make Predictions}
\usage{
make_prediction(newdata, y, positive = NULL, model, scale = FALSE, auc = FALSE)
}
\arguments{
\item{newdata}{A data frame. The test dataset.}

\item{y}{A character string. The name of the target variable.}

\item{positive}{A character string. The positive class for the target variable if \code{y} is a factor. Default: \code{NULL}.}

\item{model}{A model object. The trained model returned by \code{caret::train}.}

\item{scale}{A logical value. If \code{TRUE}, scales the variable importance values to between 0-100. Default: \code{FALSE}.}

\item{auc}{A logical value. If \code{TRUE}, calculates the area under the ROC curve (AUC) for classification models. Default: \code{FALSE}.}
}
\value{
A list containing predictions, class probabilities, and corresponding plots.
\itemize{
\item For classification models, class probabilities and ROC curves are included.
\item For regression models, predictions and residuals vs. predicted plots are included.
}
}
\description{
\code{make_prediction} makes predictions. It returns a list containing a data frame of predictions or class probabilities, and corresponding plots.
}
\details{
This function uses \code{caret}, \code{ggplot2}, and \code{plotROC} for calculations and plotting.
}
\examples{
\dontrun{
  # Example for classification model
  data("septoriaDurumWC")
  knn.mod <- caret::train(ST_S ~ ., data = septoriaDurumWC, method = "knn")
  testdata <- septoriaDurumWC
  knn.pred <- make_prediction(newdata = testdata, y = "ST_S", positive = "R", model = knn.mod)
  print(knn.pred)

  # Example with SVM and ROC curve
  svm.mod <- caret::train(ST_S ~ ., data = septoriaDurumWC, method = "svmLinear2",
   metric = "Accuracy")
  testdata <- septoriaDurumWC
  svm.pred <- make_prediction(newdata = testdata, y = "ST_S", positive = "R", model = svm.mod,
   auc = TRUE)
  print(svm.pred)
}
}
\seealso{
\code{\link[caret]{predict.train}}
}
\author{
Zakaria Kehel, Bancy Ngatia, Khadija Aziz, Zainab Azough, Chafik Analy, Khadija Aouzal
}
