% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{septoriaDurumWC}
\alias{septoriaDurumWC}
\title{septoriaDurumWC}
\format{
200 sites from durum wheat collection and their daily climatic data and evaluation for Septoria Tritici.
}
\usage{
data(septoriaDurumWC)
}
\description{
A sample data including daily data for 4 climatic variables (tmin, tmax, precipitation and relative humidity) and evaluation for Septoria Tritici
}
\examples{
if(interactive()){
 #Load durum wheat data with septoria scores and climatic variables obtained from WorldClim database
 data(septoriaDurumWC)
}
}
\keyword{datasets}
