% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{FIGS}
\alias{FIGS}
\title{FIGS}
\format{
A data frame with 201 rows and 15 variables
}
\usage{
data("FIGS")
}
\description{
FIGS subset for wheat sodicity resistance
constructed using the FAO harmonized world soil database HWSD
}
\examples{
if(interactive()){
 data("FIGS")
}
}
\references{
\href{https://openknowledge.fao.org/server/api/core/bitstreams/2f45beea-e165-4264-951e-20b7c1da65c0/content}{HWSD}
}
\keyword{datasets}
