% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getData.R
\name{getAccessions}
\alias{getAccessions}
\title{Retrieving ICARDA Genebank Accession Passport Data.}
\usage{
getAccessions(
  crop = "",
  ori = NULL,
  IG = "",
  doi = lifecycle::deprecated(),
  taxon = lifecycle::deprecated(),
  collectionYear = lifecycle::deprecated(),
  coor = FALSE,
  available = FALSE,
  other_id = FALSE
)
}
\arguments{
\item{crop}{Crop name. Default: "".}

\item{ori}{Country of origin (ISO 3166-1 alpha-3 code). Default: NULL.}

\item{IG}{List of accession numbers. Default: "".}

\item{doi}{\text{\link{Deprecated}} No longer used as of version 2.0.0.}

\item{taxon}{\text{\link{Deprecated}} No longer used as of version 2.0.0.}

\item{collectionYear}{\text{\link{Deprecated}} No longer used as of version 2.0.0.}

\item{coor}{If \code{TRUE}, returns only georeferenced accessions. Default: FALSE.}

\item{available}{If \code{TRUE}, returns only available accessions for distribution, Default: FALSE.}

\item{other_id}{If \code{TRUE}, returns other IDs associated with accessions. Default: FALSE.}
}
\value{
Data frame of accession passport data by crop or accession numbers.
}
\description{
Get ICARDA Genebank passport data by crop or accession numbers.
}
\details{
Available crops can be retrieved using \code{\link[icardaFIGSr]{getCrops}}.
}
\examples{
\dontrun{
 # Obtain accession data for durum wheat
 durum <- getAccessions(crop = 'Durum wheat', coor = TRUE)
 }
}
\author{
Khadija Aouzal, Amal Ibnelhobyb, Zakaria Kehel, Fawzy Nawar
}
