% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getData.R
\name{getAccessions}
\alias{getAccessions}
\title{Getting Accession Data from ICARDA's Genebank Database.}
\usage{
getAccessions(
  crop = "",
  ori = NULL,
  IG = "",
  doi = FALSE,
  taxon = FALSE,
  collectionYear = FALSE,
  coor = FALSE,
  available = FALSE,
  other_id = FALSE
)
}
\arguments{
\item{crop}{character. Crop for which to get accession data. See section 'Details' for available crops or use \code{\link[icardaFIGSr]{getCrops}} function. Default: "".}

\item{ori}{string. Country of origin using the ISO 3166-1 alpha-3 country codes. Default: NULL.}

\item{IG}{integer. List of accession numbers. Default: "".}

\item{doi}{boolean. If \code{TRUE} , the function will return the digital object identifiers DOI for the accessions. Default: FALSE.}

\item{taxon}{boolean. If \code{TRUE}, the function will return the taxon information of the accessions. Default: FALSE.}

\item{collectionYear}{boolean. If \code{TRUE}, the function will return the year of the collecting mission. Default: FALSE.}

\item{coor}{boolean. If \code{TRUE}, returns only georeferenced accessions. Default: FALSE.}

\item{available}{boolean. If \code{TRUE}, returns only available accessions for distribution, Default: FALSE.}
}
\value{
A data frame with accession passport data for specified crop in \code{crop} from the locations in \code{ori} if specified.
}
\description{
Return a data frame with accession data for the specified crop or specified list of accession numbers (i.e. IGs).
}
\details{
The list of available crops can be fetched from ICARDA's Genebank database using \code{\link[icardaFIGSr]{getCrops}}.
}
\examples{
\dontrun{
 # Obtain accession data for durum wheat
 durum <- getAccessions(crop = 'Durum wheat', coor = TRUE)
 }
}
\author{
Khadija Aouzal, Amal Ibnelhobyb, Zakaria Kehel, Fawzy Nawar
}
