% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCrops.R, R/getData.R
\name{getCrops}
\alias{getCrops}
\title{Crops Available in ICARDA's Genebank}
\usage{
getCrops()

getCrops()
}
\value{
A list containing all crops available in ICARDA's Genebank Database.

A data frame with crop codes and names.
}
\description{
Returns a list of crops available in ICARDA's Genebank Database.

Retrieve a list of available crops in ICARDA's Genebank.
}
\examples{
\dontrun{
 # Get list of available crops
 crops <- getCrops()
 }
\dontrun{
 # Get list of available crops
 crops <- getCrops()
 }
}
\author{
Zakaria Kehel, Fawzy Nawar, Khadija Aouzal

Khadija Aouzal, Zakaria Kehel, Fawzy Nawar
}
