% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCrops.R
\name{getCrops}
\alias{getCrops}
\title{Crops Available in ICARDA's Genebank}
\usage{
getCrops()
}
\value{
A list containing all crops available in ICARDA's Genebank Documentation System.
}
\description{
this function allows to obtain a list of crops available in ICARDA's Genebank Documentation System, it returns a list with codes and names of available crops.
}
\details{
The crop codes and names are fetched from ICARDA's online server.
}
\examples{
if(interactive()){
 # Get list of available crops
 crops <- getCrops()
 }
}
\author{
Zakaria Kehel, Fawzy Nawar
}
