% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getData.R
\name{getTraits}
\alias{getTraits}
\title{Getting Traits Descriptors for a Specific Crop}
\usage{
getTraits(crop)
}
\arguments{
\item{crop}{Crop name.}
}
\value{
A data frame with traits descriptors associated with the specified crop.
}
\description{
Return a data frame with traits descriptors associated with a specific crop.
}
\examples{
\dontrun{
if(interactive()){
 # Get traits for bread wheat
 breadTraits <- getTraits(crop = 'Bread wheat')
 }
}
}
\author{
Khadija Aouzal, Amal Ibnelhobyb, Zakaria Kehel, Fawzy Nawar
}
