% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getData.R
\name{getTraitsData}
\alias{getTraitsData}
\title{Getting Trait Data for Given Accessions}
\usage{
getTraitsData(IG, traitID)
}
\arguments{
\item{IG}{Accession numbers.}

\item{traitID}{Unique identifier of trait (from \code{\link[icardaFIGSr]{getTraits}}).}
}
\value{
A data frame with scores for the trait specified in
\code{traitID} for the accessions given in \code{IG}.
}
\description{
Return a data frame with scores for a specific trait for given accessions.
}
\details{
Valid inputs for \code{traitID} can be fetched using
the \code{\link[icardaFIGSr]{getTraits}} function.
}
\examples{
\dontrun{
if(interactive()){
 # Check trait ID for septoria and get septoria data for durum wheat
 durum <- getAccessions(crop = 'Durum wheat', coor = TRUE)
 durumTraits <- getTraits(crop = 'Durum wheat')
 septoria <- getTraitsData(IG = durum$AccessionNumber, traitID = 145)
 }
}
}
\author{
Khadija Aouzal, Amal Ibnelhobyb, Zakaria Kehel, Fawzy Nawar
}
