% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_prediction.R
\name{varimpPred}
\alias{varimpPred}
\title{Variable Importance and Predictions}
\usage{
varimpPred(
  newdata,
  y,
  positive = NULL,
  model,
  scale = FALSE,
  auc = FALSE,
  predict = FALSE,
  type = "prob",
  ...
)
}
\arguments{
\item{newdata}{A data frame. The test dataset.}

\item{y}{A character string. The name of the target variable.}

\item{positive}{A character string. The positive class for the target variable if \code{y} is a factor. Default: \code{NULL}.}

\item{model}{A model object. The trained model returned by \code{caret::train}.}

\item{scale}{A logical value. If \code{TRUE}, scales the variable importance values to between 0-100. Default: \code{FALSE}.}

\item{auc}{A logical value. If \code{TRUE}, calculates the area under the ROC curve (AUC) for classification models. Default: \code{FALSE}.}

\item{predict}{A logical value. If \code{TRUE}, calculates predictions or class probabilities. Default: \code{FALSE}.}

\item{type}{A character string. The type of prediction, e.g., \code{"prob"} or \code{"raw"}. Default: \code{"prob"}.}

\item{...}{Additional arguments to be passed to \code{caret::varImp}.}
}
\value{
A list containing variable importance scores, predictions, class probabilities, and corresponding plots.
\itemize{
\item For classification models, class probabilities and ROC curves are included if \code{predict = TRUE}.
\item For regression models, predictions and residuals vs. predicted plots are included if \code{predict = TRUE}.
}
}
\description{
\code{varimpPred} calculates variable importance and makes predictions. It returns a list containing a data frame of variable importance scores, predictions or class probabilities, and corresponding plots.
}
\details{
This function uses \code{caret}, \code{ggplot2}, and \code{plotROC} for calculations and plotting. Variable importance is calculated based on the type of model.

For example, for regression models, the absolute value of the t-statistic of each parameter is used in the importance calculation.
For classification models, a variable importance score is calculated for each class (except for some tree-based methods).
See \code{caret::varImp} for details.
}
\examples{
\dontrun{
  # Example for classification model
  data("septoriaDurumWC")
  knn.mod <- caret::train(ST_S ~ ., data = septoriaDurumWC, method = "knn")
  testdata <- septoriaDurumWC
  knn.varimp <- varimpPred(newdata = testdata, y = "ST_S", positive = "R", model = knn.mod)
  print(knn.varimp)

  # Example with SVM and ROC curve
  svm.mod <- caret::train(ST_S ~ ., data = septoriaDurumWC, method = "svmLinear2",
   metric = "Accuracy")
  testdata <- septoriaDurumWC
  svm.varimp <- varimpPred(newdata = testdata, y = "ST_S", positive = "R", model = svm.mod,
   auc = TRUE, predict = TRUE)
  print(svm.varimp)
}
}
\seealso{
\code{\link[caret]{varImp}}, \code{\link[caret]{predict.train}}
}
\author{
Zakaria Kehel, Bancy Ngatia, Khadija Aziz, Zainab Azough, Chafik Analy
}
